requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
plotly==5.13.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
tenacity==8.2.2
urllib3==1.26.14
